keyword_analyzer_V1/requirements.txt
streamlit
pandas
numpy
scikit-learn
plotly
matplotlib
seaborn
google-generativeai
jinaai
sentence-transformers
umap-learn
hdbscan
tqdm
python-dotenv
openpyxl
requests # <<<< DODANA BIBLIOTEKA >>>> 